+++ packages/library-base/src/commonMain/kotlin/io/realm/internal/RealmObjectCompanion.kt
@@ -24,8 +24,9 @@ import kotlin.reflect.KMutableProperty1
 @Suppress("VariableNaming")
 public interface RealmObjectCompanion {
     public val `io_realm_kotlin_className`: String
-    public val `io_realm_kotlin_fields`: List<KMutableProperty1<*, *>>?
+    public val `io_realm_kotlin_fields`: Map<String, KMutableProperty1<*, *>>
     public val `io_realm_kotlin_primaryKey`: KMutableProperty1<*, *>?
+    public val `io_realm_kotlin_isEmbedded`: Boolean
     public fun `io_realm_kotlin_schema`(): RealmClassImpl
     public fun `io_realm_kotlin_newInstance`(): Any
 }