+++ packages/plugin-compiler/src/main/kotlin/io/realm/compiler/IrUtils.kt
@@ -16,12 +16,16 @@
 
 package io.realm.compiler
 
+import io.realm.compiler.FqNames.BASE_REALM_OBJECT_INTERFACE
+import io.realm.compiler.FqNames.EMBEDDED_OBJECT_INTERFACE
 import io.realm.compiler.FqNames.KOTLIN_COLLECTIONS_LISTOF
 import org.jetbrains.kotlin.backend.common.extensions.IrPluginContext
 import org.jetbrains.kotlin.backend.common.ir.copyTo
 import org.jetbrains.kotlin.backend.common.lower.DeclarationIrBuilder
 import org.jetbrains.kotlin.cli.common.messages.CompilerMessageLocationWithRange
 import org.jetbrains.kotlin.cli.common.messages.CompilerMessageSourceLocation
+import org.jetbrains.kotlin.com.intellij.psi.PsiElement
+import org.jetbrains.kotlin.com.intellij.psi.PsiElementVisitor
 import org.jetbrains.kotlin.descriptors.ClassDescriptor
 import org.jetbrains.kotlin.descriptors.DescriptorVisibilities
 import org.jetbrains.kotlin.descriptors.Modality
@@ -78,10 +82,10 @@ import org.jetbrains.kotlin.ir.util.hasAnnotation
 import org.jetbrains.kotlin.ir.util.isVararg
 import org.jetbrains.kotlin.ir.util.nameForIrSerialization
 import org.jetbrains.kotlin.ir.util.properties
+import org.jetbrains.kotlin.js.resolve.diagnostics.findPsi
 import org.jetbrains.kotlin.name.FqName
 import org.jetbrains.kotlin.name.Name
-import org.jetbrains.kotlin.resolve.descriptorUtil.fqNameSafe
-import org.jetbrains.kotlin.resolve.descriptorUtil.getSuperInterfaces
+import org.jetbrains.kotlin.psi.stubs.elements.KtStubElementTypes.SUPER_TYPE_LIST
 import java.util.function.Predicate
 
 // Somehow addSetter was removed from the IrProperty in https://github.com/JetBrains/kotlin/commit/d1dc938a5d7331ba43fcbb8ce53c3e17ef76a22a#diff-2726c3747ace0a1c93ad82365cf3ff18L114
@@ -105,22 +109,45 @@ fun IrPluginContext.blockBody(
     DeclarationIrBuilder(this, symbol).irBlockBody { block() }
 
 val ClassDescriptor.isRealmObjectCompanion
-    get() = isCompanionObject && (containingDeclaration as ClassDescriptor).hasRealmModelInterface
+    get() = isCompanionObject && (containingDeclaration as ClassDescriptor).isBaseRealmObject
 
-val ClassDescriptor.hasRealmModelInterface
-    get() = getSuperInterfaces().firstOrNull { it.fqNameSafe == FqNames.REALM_MODEL_INTERFACE } != null
+val realmObjectInterfaces = setOf(FqNames.REALM_OBJECT_INTERFACE, EMBEDDED_OBJECT_INTERFACE)
+
+inline fun ClassDescriptor.hasInterfacePsi(interfaces: Set<String>): Boolean {
+    // Using PSI to find super types to avoid cyclic reference (see https://github.com/realm/realm-kotlin/issues/339)
+    var hasRealmObjectAsSuperType = false
+    this.findPsi()?.acceptChildren(object : PsiElementVisitor() {
+        override fun visitElement(element: PsiElement) {
+            if (element.node.elementType == SUPER_TYPE_LIST) {
+                hasRealmObjectAsSuperType = element.node.text.findAnyOf(interfaces) != null
+            }
+        }
+    })
+
+    return hasRealmObjectAsSuperType
+}
+
+val realmObjectPsiNames = setOf("RealmObject", "io.realm.RealmObject")
+val embeddedRealmObjectPsiNames = setOf("EmbeddedRealmObject", "io.realm.EmbeddedRealmObject")
+val ClassDescriptor.isRealmObject: Boolean
+    get() = this.hasInterfacePsi(realmObjectPsiNames)
+val ClassDescriptor.isEmbeddedRealmObject: Boolean
+    get() = this.hasInterfacePsi(embeddedRealmObjectPsiNames)
+val ClassDescriptor.isBaseRealmObject: Boolean
+    get() = this.hasInterfacePsi(realmObjectPsiNames + embeddedRealmObjectPsiNames)
 
 fun IrMutableAnnotationContainer.hasAnnotation(annotation: FqName): Boolean {
     return annotations.hasAnnotation(annotation)
 }
 
-val IrMutableAnnotationContainer.isRealmModuleAnnotated
-    get() = annotations.hasAnnotation(FqNames.REALM_MODULE_ANNOTATION)
+val IrClass.isBaseRealmObject
+    get() = superTypes.any { it.classFqName in realmObjectInterfaces }
 
-val IrClass.hasRealmModelInterface
-    get() = superTypes.firstOrNull {
-        it.classFqName?.equals(FqNames.REALM_MODEL_INTERFACE) ?: false
-    } != null
+val IrClass.isRealmObject
+    get() = superTypes.any { it.classFqName == BASE_REALM_OBJECT_INTERFACE }
+
+val IrClass.isEmbeddedRealmObject: Boolean
+    get() = superTypes.any { it.classFqName == EMBEDDED_OBJECT_INTERFACE }
 
 internal fun IrFunctionBuilder.at(startOffset: Int, endOffset: Int) = also {
     this.startOffset = startOffset
@@ -163,7 +190,7 @@ internal fun IrPluginContext.lookupClassOrThrow(name: FqName): IrClass {
 
 internal fun IrPluginContext.lookupConstructorInClass(
     fqName: FqName,
-    filter: (ctor: IrConstructorSymbol) -> Boolean
+    filter: (ctor: IrConstructorSymbol) -> Boolean = { true }
 ): IrConstructorSymbol {
     return referenceConstructors(fqName).first {
         filter(it)