+++ test/base/src/androidTest/kotlin/io/realm/test/shared/RealmListTests.kt
@@ -24,10 +24,12 @@ import io.realm.RealmInstant
 import io.realm.RealmList
 import io.realm.RealmObject
 import io.realm.RealmResults
+import io.realm.entities.Sample
 import io.realm.entities.list.Level1
 import io.realm.entities.list.Level2
 import io.realm.entities.list.Level3
 import io.realm.entities.list.RealmListContainer
+import io.realm.entities.list.listTestSchema
 import io.realm.query
 import io.realm.query.find
 import io.realm.realmListOf
@@ -61,7 +63,7 @@ class RealmListTests {
     fun setup() {
         tmpDir = PlatformUtils.createTempDir()
         val configuration = RealmConfiguration.Builder(
-            schema = setOf(RealmListContainer::class, Level1::class, Level2::class, Level3::class)
+            schema = listTestSchema + setOf(Level1::class, Level2::class, Level3::class, Sample::class)
         ).directory(tmpDir).build()
         realm = Realm.open(configuration)
     }
@@ -287,14 +289,63 @@ class RealmListTests {
     }
 
     @Test
-    @Ignore // https://github.com/realm/realm-kotlin/issues/808
-    fun addAll_duplicateObject() {
+    fun add_detectsDuplicates() {
+        val leaf = Sample().apply { intField = 1 }
+        val child = Sample().apply {
+            intField = 2
+            nullableObject = leaf
+            objectListField = realmListOf(leaf, leaf)
+        }
+        realm.writeBlocking {
+            copyToRealm(Sample()).apply {
+                objectListField.add(child)
+            }
+        }
+        assertEquals(3, realm.query<Sample>().find().size)
+    }
+
+    @Test
+    fun addWithIndex_detectsDuplicates() {
+        val leaf = Sample().apply { intField = 1 }
+        val child = Sample().apply {
+            intField = 2
+            nullableObject = leaf
+            objectListField = realmListOf(leaf, leaf)
+        }
+        realm.writeBlocking {
+            copyToRealm(Sample()).apply {
+                objectListField.add(0, child)
+            }
+        }
+        assertEquals(3, realm.query<Sample>().find().size)
+    }
+
+    @Test
+    fun addAll_detectsDuplicates() {
         val child = RealmListContainer()
         val parent = RealmListContainer()
         realm.writeBlocking {
             copyToRealm(parent).apply { objectListField.addAll(listOf(child, child)) }
         }
         assertEquals(2, realm.query<RealmListContainer>().find().size)
+    }
+
+    @Test
+    fun set_detectsDuplicates() {
+        val leaf = Sample().apply { intField = 1 }
+        val child = Sample().apply {
+            intField = 2
+            nullableObject = leaf
+            objectListField = realmListOf(leaf, leaf)
+        }
+        realm.writeBlocking {
+            copyToRealm(Sample()).apply {
+                // Need to insert an object to be able to update it with set
+                objectListField.add(Sample())
+                objectListField.set(0, child)
+            }
+        }
+        assertEquals(4, realm.query<Sample>().find().size)
     }
 
     @Test
@@ -317,7 +368,7 @@ class RealmListTests {
     }
 
     private fun getCloseableRealm(): Realm =
-        RealmConfiguration.Builder(schema = setOf(RealmListContainer::class))
+        RealmConfiguration.Builder(schema = listTestSchema)
             .directory(tmpDir)
             .name("closeable.realm")
             .build().let {