+++ test/base/src/commonMain/kotlin/io/realm/test/util/Utils.kt
@@ -60,6 +60,14 @@ inline fun Realm.use(action: (Realm) -> Unit) {
         this.close()
     }
 }
+// Expose a try-with-resource pattern for Realms, but with support for Coroutines
+suspend fun Realm.useInContext(action: suspend (Realm) -> Unit) {
+    try {
+        action(this)
+    } finally {
+        this.close()
+    }
+}
 
 // Convert Kotlinx-datatime Instant to RealmInstant
 fun Instant.toRealmInstant(): RealmInstant {